Refactor: 일반 사용자 레이아웃을 (content) 라우트 그룹으로 분리
- RootLayout에서 Header/Hero/Footer/BottomNav, Kakao 스크립트, scroll-reveal observer 제거하고 html·body·폰트·세션 컨텍스트만 유지
- (content)/layout.tsx가 공용 chrome과 관련 스크립트·data-scrolled 초기화를 담당
- app/page.tsx → app/(content)/page.tsx 로 이동해 URL 경로 유지
