+++ src/app/(content)/page.tsx
@@ -1,4 +1,4 @@
-import { Banner, QuickAccess, RecentSermons, NewHere, FeedSection } from './_component/home';
+import { Banner, QuickAccess, RecentSermons, NewHere, FeedSection } from '../_component/home';
 
 export default async function Home() {
   return (
